requirements.txt
streamlit==1.32.0
ultralytics==8.3.111
opencv-python-headless==4.9.0.80  # Use headless version for server environments
pillow==10.2.0
numpy==1.26.4
torch==2.2.2  # Add explicit PyTorch version